Flytt sigrun config til sigrun pakke
